+++ src/Core/Altinn.Platform.Authentication.Core.csproj
@@ -18,7 +18,6 @@
 
   <ItemGroup>
     <Folder Include="Configuration\" />
-    <Folder Include="Models\SystemRegisters\" />
     <Folder Include="Services\Interfaces\" />
   </ItemGroup>
