+++ src/Core/Models/Rights/AttributePair.cs
