+++ src/Core/Models/Rights/Right.cs
