+++ src/Core/Models/SystemRegisters/MaskinPortenClientInfo.cs
