+++ src/Core/Models/SystemRegisters/RegisterSystemResponse.cs
@@ -1,28 +1,45 @@
 ﻿using Altinn.Platform.Authentication.Core.Models;
 using System.ComponentModel.DataAnnotations;
 using System.Diagnostics.CodeAnalysis;
+using System.Text.Json.Serialization;
 
 #nullable enable
 namespace Altinn.Platform.Authentication.Core.SystemRegister.Models
 {
     /// <summary>
-    /// Model for the System Register
+    /// Model for the response of a registered system
     /// A Registered System is a product supplied by a System Vendor,
     /// it may need Rights to use or acccess Resources at a Service Provider.
+    /// This complete Response model is used by the Vendors' API
     /// </summary>
     [ExcludeFromCodeCoverage]
-    public class RegisterSystemRequest
+    public class RegisterSystemResponse
     {
         /// <summary>
-        /// The unique Id for this product, in human-readable string format.
-        /// The id is in the format of system_vendor_name_plus_name_chosen_by_them.
+        /// The primary key to store and edit the Registered System.
+        /// Used as a Foreign Key in the System User table.
         /// 
-        /// An Optimistic Concurrency pattern to create new System Ids is used,
-        /// where the Id of the product is prefixed with the SystemVendor to help with uniqueness.
+        /// The SystemInternalId never changes throughout the lifecycle 
+        /// of the entry.
+        /// </summary>
+        [Required]
+        public Guid SystemInternalId { get; set; } = Guid.Empty;
+
+        /// <summary>
+        /// A unique External Id for this System, in human-readable string format.
+        /// The id could be in the format of system_vendor_name_plus_name_chosen_by_them.
+        /// 
+        /// An Optimistic Concurrency pattern to create new System Ids could be used,
+        /// where the Id of the System is prefixed with the SystemVendor.
         /// 
         /// When the SystemVendor tries to register a new system (product), 
         /// they should be aware of their own previous system names 
         /// when giving the new system it's id.
+        /// 
+        /// The Registered Systems are stored using a unique system_internal_id
+        /// as the primary key, which allows System Vendors to reuse the SystemId 
+        /// if creating new systems to replace the existing one. The old system's
+        /// name should be changed prior to be reused, since it must be unique.
         /// </summary>
         [Required]
         public string SystemId { get; set; } = string.Empty;
@@ -34,10 +51,10 @@ namespace Altinn.Platform.Authentication.Core.SystemRegister.Models
         public string SystemVendorOrgNumber { get; set; }
 
         /// <summary>
-        /// A short name of the product, used when displaying to the user
+        /// Organization number of the system Vendor that offers the product (system)
         /// </summary>
         [Required]
-        public string SystemName { get; set; } = string.Empty;
+        public string SystemVendorOrgName { get; set; } = string.Empty;
 
         /// <summary>
         /// A short name of the product, used when displaying to the user
@@ -48,7 +65,6 @@ namespace Altinn.Platform.Authentication.Core.SystemRegister.Models
         /// A short description of the product, used when displaying to the user
         /// </summary>
         public required IDictionary<string, string> Description { get; set; }
-
 
         /// <summary>
         /// The array of Rights versus System Provider's Resources needed to use this Registered System
@@ -63,13 +79,18 @@ namespace Altinn.Platform.Authentication.Core.SystemRegister.Models
         /// <summary>
         /// The client Id
         /// </summary>
-        [Required]
-        public List<String> ClientId { get; set; } = [];
+        [JsonIgnore]
+        public List<string> ClientId { get; set; } = [];
 
         /// <summary>
         /// Registered systems can be set to false to hide it from the user interface.
         /// This is used when the vendor does not want the user to create system users for specific systems
         /// </summary>
         public bool IsVisible { get; set; } = true;
+
+        /// <summary>
+        /// White listing of redirect urls
+        /// </summary>
+        public List<Uri> AllowedRedirectUrls { get; set; } = [];
     }
 }
